+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/addons/AddonPopupBaseFragment.kt
@@ -121,6 +121,7 @@ abstract class AddonPopupBaseFragment : Fragment(), EngineSession.Observer, User
                                 activity = requireActivity(),
                                 filename = filename.value,
                                 contentSize = contentSize.value,
+                                fileSizeFormatter = requireComponents.core.fileSizeFormatter,
                                 positiveButtonAction = positiveAction.value,
                                 negativeButtonAction = negativeAction.value,
                             ).onDismiss {
@@ -334,6 +335,7 @@ abstract class AddonPopupBaseFragment : Fragment(), EngineSession.Observer, User
             } else {
                 val dynamicDownloadDialog = DynamicDownloadDialog(
                     context = safeContext,
+                    fileSizeFormatter = requireComponents.core.fileSizeFormatter,
                     downloadState = downloadState,
                     didFail = downloadJobStatus == DownloadState.Status.FAILED,
                     tryAgain = tryAgain,