+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/browser/DownloadUtils.kt
@@ -9,6 +9,7 @@ import mozilla.components.browser.state.state.content.DownloadState.Status
 import mozilla.components.concept.toolbar.ScrollableToolbar
 import mozilla.components.feature.downloads.AbstractFetchDownloadService
 import org.mozilla.fenix.downloads.dialog.DynamicDownloadDialog
+import org.mozilla.fenix.ext.requireComponents
 
 internal fun BaseBrowserFragment.handleOnDownloadFinished(
     downloadState: DownloadState,
@@ -39,6 +40,7 @@ internal fun BaseBrowserFragment.handleOnDownloadFinished(
 
             val dynamicDownloadDialog = DynamicDownloadDialog(
                 context = safeContext,
+                fileSizeFormatter = requireComponents.core.fileSizeFormatter,
                 downloadState = downloadState,
                 didFail = downloadJobStatus == Status.FAILED,
                 tryAgain = tryAgain,