+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/downloads/listscreen/DownloadFragment.kt
@@ -62,6 +62,7 @@ class DownloadFragment : ComposeFragment(), UserInteractionHandler, MenuProvider
             middleware = listOf(
                 DownloadUIMapperMiddleware(
                     browserStore = requireComponents.core.store,
+                    fileSizeFormatter = requireComponents.core.fileSizeFormatter,
                     scope = viewModelScope,
                 ),
             ),